[リファクタリング] GameSurviceを通してWebSocketへメッセージを送るように
GameServerが持つべき機能ではなかったため移動させた
